requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0,<4.0.0
httpx>=0.27.0
cryptography>=42.0.0
python-multipart>=0.0.9
pytest>=8.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0
email-validator>=2.0.0
